requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
requests
openpyxl
pillow
datetime